+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.0.0rc3 (unreleased)
 ---------------------
 
+- #1727 Cleanup Controlpanel Items
 - #1726 Content rules control panel templates styled
 - #1724 Purge stale metadata and indexes from analysis catalog
 - #1720 Fix UnicodeDecodeError for Instrument Import Log View